Set default precision to Float32

+++ src/InterfaceAdvection.jl
@@ -36,7 +36,7 @@ include("metrics.jl")
                         λμ=1e-2,λρ=1e-3,η=0,
                         InterfaceSDF::Function=(x) -> -5-x[1],
                         uλ=nothing, exitBC=false, body::AbstractBody=NoBody(),
-                        T=Float64, mem=Array)
+                        T=Float32, mem=Array)
 
 Constructor for a WaterLily.jl two phase simulation, which is identical to the original on with some additional properties for multiphase flow:
 
@@ -75,7 +75,7 @@ mutable struct TwoPhaseSimulation <: AbstractSimulation
                         λμ=1e-2,λρ=1e-3,η=0,
                         InterfaceSDF::Function=(x) -> -5-x[1],
                         uλ=nothing, exitBC=false, body::AbstractBody=NoBody(),
-                        T=Float64, mem=Array) where N 
+                        T=Float32, mem=Array) where N 
         # same as waterlily
         @assert !(isnothing(U) && isa(uBC,Function)) "`U` (velocity scale) must be specified if boundary conditions `uBC` is a `Function`"
         isnothing(U) && (U = √sum(abs2,uBC))
